+++ src/stats.h
@@ -2,13 +2,13 @@
 #define STATS_H
 
 typedef struct {
-    int samples;
-    int index;
-    int limit;
+    uint64_t samples;
+    uint64_t index;
+    uint64_t limit;
     uint64_t data[];
 } stats;
 
-stats *stats_alloc(int);
+stats *stats_alloc(uint64_t);
 void stats_free(stats *);
 void stats_record(stats *, uint64_t);
 uint64_t stats_min(stats *);
